+++ demos/chinook/src/main/java/is/codion/demos/chinook/model/PlaylistTrackEditModel.java
@@ -25,8 +25,6 @@ import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.swing.framework.model.SwingEntityEditModel;
 
-import java.util.List;
-
 public final class PlaylistTrackEditModel extends SwingEntityEditModel {
 
 	public PlaylistTrackEditModel(EntityConnectionProvider connectionProvider) {
@@ -39,8 +37,6 @@ public final class PlaylistTrackEditModel extends SwingEntityEditModel {
 
 	private void excludePlaylistTracks(Entity playlist) {
 		foreignKeySearchModel(PlaylistTrack.TRACK_FK).condition().set(() -> playlist == null ? null :
-						Track.NOT_IN_PLAYLIST.get(
-										List.of(Playlist.ID),
-										List.of(playlist.get(Playlist.ID))));
+						Track.NOT_IN_PLAYLIST.get(Playlist.ID, playlist.get(Playlist.ID)));
 	}
 }